+++ cheqd/cheqd/did/tests/test_manager.py
@@ -16,7 +16,7 @@ from .mocks import (
 )
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_create(mock_registrar_instance, profile):
     # Arrange
@@ -60,7 +60,7 @@ async def test_create(mock_registrar_instance, profile):
     )
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_create_with_seed(mock_registrar_instance, profile):
     # Arrange
@@ -78,7 +78,7 @@ async def test_create_with_seed(mock_registrar_instance, profile):
     assert response["didDocument"]["MOCK_KEY"] == "MOCK_VALUE"
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_create_with_insecure_seed(mock_registrar_instance, profile):
     # Arrange
@@ -96,7 +96,7 @@ async def test_create_with_insecure_seed(mock_registrar_instance, profile):
     assert str(e.value) == "Insecure seed is not allowed"
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_create_with_invalid_did_document(
     mock_registrar_instance,
@@ -118,7 +118,7 @@ async def test_create_with_invalid_did_document(
     assert str(e.value) == "Error constructing DID Document"
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_create_with_signing_failure(
     mock_registrar_instance,
@@ -140,7 +140,7 @@ async def test_create_with_signing_failure(
     assert str(e.value) == "No signing requests available for create."
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_create_with_network_failure(
     mock_registrar_instance,
@@ -162,7 +162,7 @@ async def test_create_with_network_failure(
     assert str(e.value) == "Error registering DID Network failure"
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_create_not_finished(
     mock_registrar_instance,
@@ -184,8 +184,8 @@ async def test_create_not_finished(
     assert str(e.value) == "Error registering DID Not finished"
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_update(
     mock_registrar_instance, mock_resolver_instance, profile, did, did_doc
@@ -227,8 +227,8 @@ async def test_update(
     )
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_update_with_did_deactivated(
     mock_registrar_instance, mock_resolver_instance, profile, did, did_doc
@@ -252,8 +252,8 @@ async def test_update_with_did_deactivated(
     assert str(e.value) == "DID is already deactivated or not found."
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_update_with_signing_failure(
     mock_registrar_instance, mock_resolver_instance, profile, did, did_doc
@@ -278,8 +278,8 @@ async def test_update_with_signing_failure(
     assert str(e.value) == "No signing requests available for update."
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_update_with_network_failure(
     mock_registrar_instance, mock_resolver_instance, profile, did, did_doc
@@ -304,8 +304,8 @@ async def test_update_with_network_failure(
     assert str(e.value) == "Error updating DID Network failure"
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_update_not_finished(
     mock_registrar_instance, mock_resolver_instance, profile, did, did_doc
@@ -333,8 +333,8 @@ async def test_update_not_finished(
     )
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_deactivate(mock_registrar_instance, mock_resolver_instance, profile, did):
     # Arrange
@@ -369,8 +369,8 @@ async def test_deactivate(mock_registrar_instance, mock_resolver_instance, profi
     )
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_deactivate_with_did_deactivated(
     mock_registrar_instance, mock_resolver_instance, profile, did
@@ -394,8 +394,8 @@ async def test_deactivate_with_did_deactivated(
     assert str(e.value) == "DID is already deactivated or not found."
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_deactivate_with_signing_failure(
     mock_registrar_instance, mock_resolver_instance, profile, did
@@ -420,8 +420,8 @@ async def test_deactivate_with_signing_failure(
     assert str(e.value) == "No signing requests available for update."
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_deactivate_with_network_failure(
     mock_registrar_instance, mock_resolver_instance, profile, did
@@ -446,8 +446,8 @@ async def test_deactivate_with_network_failure(
     assert str(e.value) == "Error deactivating DID Network failure"
 
 
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDResolver")
-@patch("cheqd.cheqd.v1_0.did.manager.CheqdDIDRegistrar")
+@patch("cheqd.cheqd.did.manager.CheqdDIDResolver")
+@patch("cheqd.cheqd.did.manager.CheqdDIDRegistrar")
 @pytest.mark.asyncio
 async def test_deactivate_not_finished(
     mock_registrar_instance, mock_resolver_instance, profile, did
