+++ cheqd/cheqd/did/tests/test_registrar.py
