+++ cheqd/integration/run_integration_tests.sh
@@ -2,4 +2,4 @@
 
 docker compose build
 docker compose run tests
-docker compose down -v  # Clean up
+docker compose down --remove-orphans -v  # Clean up
